telegram-service/requirements.txt
telethon==1.36.0
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
aiofiles==24.1.0

